+++ src/controllers/keywordController.js
@@ -1,5 +1,5 @@
 const request = require("request");
-const keywordService = require("../service/keywordService");
+const keywordService = require("../services/keywordService");
 require("dotenv").config();
 
 const keywordAnalyze = async (req, res) => {
